+++ src/ua.rs
@@ -12,7 +12,9 @@ mod certificate_verification;
 mod client;
 mod client_config;
 mod continuation_point;
+mod data_type_definition;
 mod data_types;
+mod duration;
 mod event_id;
 mod event_notifier_type;
 mod key_value_map;
@@ -45,7 +47,9 @@ pub use self::{
     certificate_verification::CertificateVerification,
     client::{Client, ClientState},
     continuation_point::ContinuationPoint,
+    data_type_definition::DataTypeDefinition,
     data_types::*,
+    duration::Duration,
     event_id::EventId,
     event_notifier_type::EventNotifierType,
     key_value_map::KeyValueMap,