+++ src/ua/data_types/structure_definition.rs
@@ -1,0 +1,10 @@
+use crate::{ua, DataType};
+
+crate::data_type!(StructureDefinition);
+
+impl StructureDefinition {
+    #[must_use]
+    pub fn base_data_type(&self) -> &ua::NodeId {
+        ua::NodeId::raw_ref(&self.0.baseDataType)
+    }
+}